backend/requirements.txt
pydantic-settings
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-dotenv
fugashi
unidic-lite
python-multipart
pytest